refactored middleware implementation
