+++ example/main.go
@@ -8,17 +8,24 @@ import (
 	"github.com/jsphbtst/go-express"
 )
 
-func SampleCustomMiddleware(w http.ResponseWriter, r *http.Request) {
-	if r.URL.Path == "/redirect" && r.Method == http.MethodGet {
-		log.Println("REDIRECT URL IDK MAN")
+func SampleCustomMiddleware(config express.Config) express.Handler {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/redirect" && r.Method == http.MethodGet {
+			log.Println("REDIRECT URL IDK MAN")
+		}
 	}
 }
 
 func main() {
 	app := express.New()
-	app.Use(express.Cors)
-	app.Use(express.LogPathAccess)
-	app.Use(SampleCustomMiddleware)
+
+	app.Use(express.Cors(
+		express.Config{
+			"origin": "*",
+		},
+	))
+	app.Use(express.LogPathAccess(nil))
+	app.Use(SampleCustomMiddleware(nil))
 
 	app.Get("/", func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
