+++ middlewares.go
@@ -5,12 +5,20 @@ import (
 	"net/http"
 )
 
-func LogPathAccess(w http.ResponseWriter, r *http.Request) {
-	path := r.URL.Path
-	log.Printf("%s `%s`\n", r.Method, path)
+func LogPathAccess(config Config) Handler {
+	return func(w http.ResponseWriter, r *http.Request) {
+		path := r.URL.Path
+		log.Printf("%s `%s`\n", r.Method, path)
+	}
 }
 
-// TODO: allow for origin to be changed lol
-func Cors(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Access-Control-Allow-Origin", "*")
+func Cors(config Config) Handler {
+	return func(w http.ResponseWriter, r *http.Request) {
+		origin, ok := config["origin"]
+		if !ok {
+			w.Header().Set("Access-Control-Allow-Origin", "*")
+		} else {
+			w.Header().Set("Access-Control-Allow-Origin", origin)
+		}
+	}
 }
